feat(nvim-nightly): keep lua config in another repo (read commit msg)
Keeping this directory as a direct translation of my stable config so I
can use this to replace vim plugins with lua config and check their
performance as compared to my original/stable config.

At this point creativenull/nvim-config takes ~120ms to startup, even tho
majority of the plugins are in lua and should be theoratically faster
than my stable config.

However, that was not the case with my config. My stable config takes
~60ms (nearly twice as fast compared to the lua config), now this could
also be that the lua config is structured mostly on the `lua/` directory
and could be running unnecessary code that is not need on init. But from
my observations the culprit is always setting the colorscheme.

More investigation will be needed but for now, I'll be slowly translating
my stable config 1-to-1 to this directory and check how to optimize
startup.

Also to note, this does not mean that lua itself is slow, IMO any startup
< 500ms is a good startup and a user should not need to wait more than
that to start using vim. I just want to see how fast I can get my vim
starting up :)